+++ lib/plugins/aws/package/compile/events/sns.js
