+++ test/unit/lib/plugins/aws/package/compile/events/sns.test.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../../../../../lib/plugins/aws/provider/awsProvider');
-const AwsCompileSNSEvents = require('../../../../../../../../../lib/plugins/aws/package/compile/events/sns/index');
-const Serverless = require('../../../../../../../../../lib/Serverless');
+const AwsProvider = require('../../../../../../../../lib/plugins/aws/provider/awsProvider');
+const AwsCompileSNSEvents = require('../../../../../../../../lib/plugins/aws/package/compile/events/sns');
+const Serverless = require('../../../../../../../../lib/Serverless');
 
 describe('AwsCompileSNSEvents', () => {
   let serverless;
